+++ node/src/types/kernel/api/storage/setLocalStorage/index.d.ts
@@ -17,9 +17,8 @@
 
 //#region content
 /* directories */
-export * as getLocalStorage from "./getLocalStorage";
-export * as getRecentDocs from "./getRecentDocs";
-export * as setLocalStorage from "./setLocalStorage";
 
 /* files */
+export * from "./payload";
+export * from "./response";
 //#endregion content