+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - 添加 API `/api/system/getConf` | Add API `/api/system/getConf`
 - 完善 API `/api/system/getConf` | Perfect API `/api/system/getConf`
 - 添加 API `/api/storage/getLocalStorage` | Add API `/api/storage/getLocalStorage`
+- 添加 API `/api/storage/setLocalStorage` | Add API `/api/storage/setLocalStorage`
 
 ## v0.3.2 / 2023-09-19